fix: Resolve TypeScript and C# build errors
TypeScript fixes:
- Add zodResolver type assertion to inventory/[id], inventory/new, settings/specimens
- Update cycles/new and inventory/new to use new SpecimenMultiSelect interface
  (selectedItems: SpecimenSelection[] instead of deprecated props)

C# fixes:
- Cast totalRuntimeHours from long to int in CycleService.cs and PostService.cs
  (safe cast - runtime hours won't exceed int.MaxValue in practice)

+++ src/api/MyUglyRocks.Core/Services/CycleService.cs
@@ -300,7 +300,7 @@ public class CycleService : ICycleService
             stageRunSummaries,
             specimens,
             elapsedDays,
-            totalRuntimeHours,
+            (int)totalRuntimeHours,
             completedStagesCount,
             activeStageName,
             lastUpdated,
